refactor: rename app.ts to appRepository.ts for better naming consistency
- Renamed packages/lib/server/repository/app.ts to appRepository.ts
- Updated import statement in deleteCredential.test.ts
- No functional changes, only file rename and import path update
