+++ packages/features/credentials/deleteCredential.test.ts
@@ -5,7 +5,7 @@ import {
 
 import { describe, test, expect, beforeEach } from "vitest";
 
-import { AppRepository } from "@calcom/lib/server/repository/app";
+import { AppRepository } from "@calcom/lib/server/repository/appRepository";
 import { CredentialRepository } from "@calcom/lib/server/repository/credential";
 import { DestinationCalendarRepository } from "@calcom/lib/server/repository/destinationCalendar";
 import { EventTypeRepository } from "@calcom/lib/server/repository/eventTypeRepository";
